mod gt type to long tensor

+++ label_anything/data/coco.py
@@ -741,7 +741,7 @@ class CocoLVISTestDataset(CocoLVISDataset):
             "dims": dims,
         }
 
-        return data_dict, gt
+        return data_dict, gt.long()
 
 
 class LabelAnyThingOnlyImageDataset(Dataset):
